SkillOne/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.0.3
scikit-learn==1.3.2
numpy==1.26.2
networkx==3.2
gensim==4.3.1
sentence-transformers==2.2.2
pandas==2.1.1
scipy==1.11.4
